more protected variables to trait
